pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hfcx-ai-claims"
version = "1.0.0"
description = "AI-Powered Claims Processing Layer for HealthFlow HCX Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "httpx==0.27.2",

    # LangGraph / LangChain (agent orchestration)
    "langgraph==0.2.28",
    "langchain==0.3.7",
    "langchain-core==0.3.15",
    "langchain-community==0.3.5",

    # LiteLLM (unified LLM gateway)
    "litellm==1.51.1",

    # Kafka
    "aiokafka==0.11.0",
    "confluent-kafka==2.5.3",

    # Database
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "psycopg2-binary==2.9.10",

    # Redis
    "redis[hiredis]==5.1.1",
    "aioredis==2.0.1",

    # FHIR
    "fhir.resources==8.0.0",

    # Vector store (RAG)
    "chromadb==0.5.18",
    "sentence-transformers==3.2.1",

    # Healthcare ML / NLP
    "pyhealth==1.1.5",
    "scikit-learn==1.5.2",
    "xgboost==2.1.2",
    "pyod==2.0.2",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "pandas==2.2.3",

    # Auth / Security
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "cryptography==43.0.3",

    # Observability
    "opentelemetry-sdk==1.27.0",
    "opentelemetry-instrumentation-fastapi==0.48b0",
    "opentelemetry-exporter-otlp==1.27.0",
    "prometheus-client==0.21.0",

    # Logging
    "structlog==24.4.0",

    # Config
    "pydantic-settings==2.6.0",
    "python-dotenv==1.0.1",

    # Utilities
    "tenacity==9.0.0",
    "cachetools==5.5.0",
    "faker==30.8.2",
    "mimesis==18.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "testcontainers==4.8.1",
    "httpx==0.27.2",
    "ruff==0.7.1",
    "mypy==1.13.0",
    "pre-commit==4.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"
